// PetRadar-Web-UI/src/app/api/petradar/model/userCreateModel.ts
/**
 * PetRadar.Web.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RoleEnum } from './roleEnum';


export interface UserCreateModel { 
    email: string;
    password: string;
    name: string;
    lastName?: string | null;
    phoneNumber?: string | null;
    organizationName?: string | null;
    organizationAddress?: string | null;
    organizationPhone?: string | null;
    role?: RoleEnum;
}
export namespace UserCreateModel {
}