// PetRadar-Web-UI/src/app/api/petradar/model/userUpdateModel.ts
/**
 * PetRadar.Web.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RoleEnum } from './roleEnum';


export interface UserUpdateModel { 
    email?: string | null;
    password?: string | null;
    name?: string | null;
    lastName?: string | null;
    phoneNumber?: string | null;
    organizationName?: string | null;
    organizationAddress?: string | null;
    organizationPhone?: string | null;
    role?: RoleEnum;
}
export namespace UserUpdateModel {
}