// PetRadar-Web-UI/src/app/api/petradar/model/userViewModel.ts
/**
 * PetRadar.Web.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserViewModel { 
    id?: number;
    email?: string | null;
    name?: string | null;
    lastName?: string | null;
    phoneNumber?: string | null;
    profilePhotoURL?: string | null;
    role?: string | null;
    organizationName?: string | null;
    organizationAddress?: string | null;
    organizationPhone?: string | null;
}